requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
feedparser>=6.0.11
selectolax>=0.3.21
flask>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
rapidfuzz>=3.9.0
apscheduler>=3.10.4
